+++ docs/src/content/img/react.mdx
@@ -44,7 +44,7 @@ import { Image } from "@unpic/react";
 
 ## Unpic for Next.js
 
-If you are using Next.js, you can use the `@unpic/react/next` component. This
+If you are using Next.js, you can use the `@unpic/react/nextjs` component. This
 works the same as the main `@unpic/react` component, but also supports the
 Next.js Image Optimizer for images that don't use a supported CDN. This includes
 local images, either using a path or an import.
@@ -58,10 +58,10 @@ same way:
 npm install @unpic/react
 ```
 
-Then, in your Next.js project, import the component from `@unpic/react/next`:
+Then, in your Next.js project, import the component from `@unpic/react/nextjs`:
 
 ```tsx
-import { Image } from "@unpic/react/next";
+import { Image } from "@unpic/react/nextjs";
 import logo from "../public/logo.png";
 
 // You can use both local images, and ones from image CDNs
@@ -75,6 +75,9 @@ import logo from "../public/logo.png";
   alt="Shopify product"
 />
 ```
+
+> **:warning: For versions of Next.js before 14.0.0, import from
+> `@unpic/react/next-legacy`, not `@unpic/react/nextjs`**
 
 ### Differences from `next/image`
