+++ packages/react/next-legacy.d.ts
@@ -1,0 +1,1 @@
+export * from "./dist/next-legacy.js";
