+++ packages/react/package.json
@@ -10,6 +10,8 @@
   "homepage": "https://unpic.pics/img/react",
   "files": [
     "dist",
+    "nextjs.d.ts",
+    "next-legacy.d.ts",
     "next.d.ts"
   ],
   "main": "./dist/index.js",
@@ -35,16 +37,38 @@
         "default": "./dist/next.mjs"
       }
     },
+    "./next-legacy": {
+      "require": {
+        "types": "./dist/next-legacy.d.ts",
+        "default": "./dist/next-legacy.js"
+      },
+      "import": {
+        "types": "./dist/next-legacy.d.mts",
+        "default": "./dist/next-legacy.mjs"
+      }
+    },
+    "./nextjs": {
+      "require": {
+        "types": "./dist/nextjs.d.ts",
+        "default": "./dist/nextjs.js"
+      },
+      "import": {
+        "types": "./dist/nextjs.d.mts",
+        "default": "./dist/nextjs.mjs"
+      }
+    },
     "./package.json": "./package.json"
   },
   "types": "./dist/index.d.ts",
   "scripts": {
-    "build": "tsup src/index.ts src/next.tsx --format cjs,esm --dts --clean",
-    "prepublishOnly": "pnpm build",
+    "build": "tsup src/index.ts src/next.tsx src/nextjs.tsx src/next-legacy.tsx --format cjs,esm --dts --clean",
+    "prepublishOnly": "pnpm build && pnpm publint && pnpm attw",
     "publint": "publint",
+    "typelint": "attw --pack .",
     "test": "vitest run"
   },
   "devDependencies": {
+    "@arethetypeswrong/cli": "^0.13.1",
     "@testing-library/react": "^14.0.0",
     "@types/node": "^20.0.0",
     "@types/react": "^18.2.34",
@@ -54,14 +78,14 @@
     "jsdom": "^22.0.0",
     "next": "^14.0.0",
     "publint": "^0.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "tsup": "^7.2.0",
     "typescript": "^5.2.2",
     "vitest": "^0.32.0"
   },
   "dependencies": {
-    "@unpic/core": "workspace:^",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "@unpic/core": "workspace:^"
   },
   "peerDependencies": {
     "next": ">=14",
